+++ src/Exception/RepositoryException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ixocreate\Database\Exception;
+
+class RepositoryException extends \Exception
+{
+
+}