+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/database-package/compare/0.2.0...develop)
+## [Unreleased](https://github.com/ixocreate/database-package/compare/0.2.1...develop)
+
+## [v0.2.1 (2019-09-12)](https://github.com/ixocreate/database-package/compare/0.2.0...0.2.1)
+### Changed
+- remove metadata loading fallback
+- add checks to DoctrineRepositoryFactory
 
 ## [v0.2.0 (2019-05-06)](https://github.com/ixocreate/database-package/compare/0.1.0...0.2.0)
 ### Changed